+++ guillotina/contrib/dyncontent/__init__.py
@@ -1,0 +1,9 @@
+from guillotina import configure
+
+
+app_settings = {}
+
+
+def includeme(root, settings):
+    configure.scan("guillotina.contrib.dyncontent.vocabularies")
+    configure.scan("guillotina.contrib.dyncontent.subscriber")